// app/src/features/presentation/types/slide.ts
export type GradientType = 'linear' | 'radial';
export type GradientColor = {
  pos: number;
  color: string;
};
export interface Gradient {
  type: GradientType;
  colors: GradientColor[];
  rotate: number;
}

export type LineStyleType = 'solid' | 'dashed' | 'dotted';

/**
 * Element shadow
 *
 * h: horizontal offset
 * v: vertical offset
 * blur: blur amount
 * color: shadow color
 */
export interface PPTElementShadow {
  h: number;
  v: number;
  blur: number;
  color: string;
}

/**
 * Element border
 *
 * style?: border style (solid or dashed)
 * width?: border width
 * color?: border color
 */
export interface PPTElementOutline {
  style?: LineStyleType;
  width?: number;
  color?: string;
}
// ...
export interface PPTTextElement extends PPTBaseElement {
  type: 'text';
  content: string;
  defaultFontName: string;
  defaultColor: string;
  outline?: PPTElementOutline;
  fill?: string;
  lineHeight?: number;
  wordSpace?: number;
  opacity?: number;
  shadow?: PPTElementShadow;
  paragraphSpace?: number;
  vertical?: boolean;
  textType?: TextType;
}
// ...
export interface PPTImageElement extends PPTBaseElement {
  type: 'image';
  fixedRatio: boolean;
  src: string;
  outline?: PPTElementOutline;
  filters?: ImageElementFilters;
  clip?: ImageElementClip;
  flipH?: boolean;
  flipV?: boolean;
  shadow?: PPTElementShadow;
  radius?: number;
  colorMask?: string;
  imageType?: ImageType;
}
// ...
export interface PPTShapeElement extends PPTBaseElement {
  type: 'shape';
  viewBox: [number, number];
  path: string;
  fixedRatio: boolean;
  fill: string;
  gradient?: Gradient;
  pattern?: string;
  outline?: PPTElementOutline;
  opacity?: number;
  flipH?: boolean;
  flipV?: boolean;
  shadow?: PPTElementShadow;
  special?: boolean;
  text?: ShapeText;
  pathFormula?: ShapePathFormulasKeys;
  keypoints?: number[];
}
// ...
export interface PPTLineElement extends Omit<PPTBaseElement, 'height' | 'rotate'> {
  type: 'line';
  start: [number, number];
  end: [number, number];
  style: LineStyleType;
  color: string;
  points: [LinePoint, LinePoint];
  shadow?: PPTElementShadow;
  broken?: [number, number];
  broken2?: [number, number];
  curve?: [number, number];
  cubic?: [[number, number], [number, number]];
}
// ...
export interface PPTChartElement extends PPTBaseElement {
  type: 'chart';
  fill?: string;
  chartType: ChartType;
  data: ChartData;
  options?: ChartOptions;
  outline?: PPTElementOutline;
  themeColors: string[];
  textColor?: string;
}
// ...
export interface PPTTableElement extends PPTBaseElement {
  type: 'table';
  outline: PPTElementOutline;
  theme?: TableTheme;
  colWidths: number[];
  cellMinHeight: number;
  data: TableCell[][];
}

/**
 * LaTeX element (formula)
 *
 * type: element type (latex)
 * latex: latex code
 * path: svg path
 * color: color
 * strokeWidth: path width
 * viewBox: SVG's viewBox attribute
 * fixedRatio: fixed shape aspect ratio
 */
export interface PPTLatexElement extends PPTBaseElement {
  type: 'latex';
  latex: string;
  path: string;
  color: string;
  strokeWidth: number;
  viewBox: [number, number];
  fixedRatio: boolean;
}

/**
 * Video element
 *
 * type: element type (video)
 * src: video address
 * autoplay: autoplay
 * poster: preview cover
 * ext: video suffix, use this field to confirm resource type when resource link is missing
 */
export interface PPTVideoElement extends PPTBaseElement {
  type: 'video';
  src: string;
  autoplay: boolean;
  poster?: string;
  ext?: string;
}

/**
 * Audio element
 *
 * type: element type (audio)
 * fixedRatio: fixed icon aspect ratio
 * color: icon color
 * loop: loop play
 * autoplay: autoplay
 * src: audio address
 * ext: audio suffix, use this field to confirm resource type when resource link is missing
 */
export interface PPTAudioElement extends PPTBaseElement {
  type: 'audio';
  fixedRatio: boolean;
  color: string;
  loop: boolean;
  autoplay: boolean;
  src: string;
  ext?: string;
}

export type PPTElement =
  | PPTTextElement
  | PPTImageElement
  | PPTShapeElement
  | PPTLineElement
  | PPTChartElement
  | PPTTableElement
  | PPTLatexElement
  | PPTVideoElement
  | PPTAudioElement;

export type AnimationType = 'in' | 'out' | 'attention';
export type AnimationTrigger = 'click' | 'meantime' | 'auto';

/**
 * Element animation
 *
 * id: animation id
 * elId: element ID
 * effect: animation effect
 * type: animation type (entry, exit, emphasis)
 * duration: animation duration
 * trigger: animation trigger method (click - single click, meantime - at the same time as the previous animation, auto - after the previous animation)
 */
export interface PPTAnimation {
  id: string;
  elId: string;
  effect: string;
  type: AnimationType;
  duration: number;
  trigger: AnimationTrigger;
}

export type SlideBackgroundType = 'solid' | 'image' | 'gradient';
export type SlideBackgroundImageSize = 'cover' | 'contain' | 'repeat';
export interface SlideBackgroundImage {
  src: string;
  size: SlideBackgroundImageSize;
}

/**
 * Slide background
 *
 * type: background type (solid, image, gradient)
 * color?: background color (solid)
 * image?: image background
 * gradientType?: gradient background
 */
export interface SlideBackground {
  type: SlideBackgroundType;
  color?: string;
  image?: SlideBackgroundImage;
  gradient?: Gradient;
}

export type TurningMode =
  | 'no'
  | 'fade'
  | 'slideX'
  | 'slideY'
  | 'random'
  | 'slideX3D'
  | 'slideY3D'
  | 'rotate'
  | 'scaleY'
  | 'scaleX'
  | 'scale'
  | 'scaleReverse';

export interface NoteReply {
  id: string;
  content: string;
  time: number;
  user: string;
}

export interface Note {
  id: string;
  content: string;
  time: number;
  user: string;
  elId?: string;
  replies?: NoteReply[];
}

export interface SectionTag {
  id: string;
  title?: string;
}

export type SlideType = 'cover' | 'contents' | 'transition' | 'content' | 'end';

/**
 * Slide page
 *
 * id: page ID
 * elements: element collection
 * notes?: comments
 * remark?: remark
 * background?: page background
 * animations?: element animation collection
 * turningMode?: turning mode
 * slideType?: page type
 */
export interface Slide {
  id: string;
  elements: PPTElement[];
  notes?: Note[];
  remark?: string;
  background?: SlideBackground;
  animations?: PPTAnimation[];
  turningMode?: TurningMode;
  sectionTag?: SectionTag;
  type?: SlideType;
}

/**
 * Slide theme
 *
 * backgroundColor: page background color
 * themeColor: theme color, used for default shape color, etc.
 * fontColor: font color
 * fontName: font
 */
export interface SlideTheme {
  backgroundColor: string | Gradient;
  themeColors: string[];
  fontColor: string;
  fontName: string;
  outline: PPTElementOutline;
  shadow: PPTElementShadow;
  titleFontName?: string;
  titleFontColor?: string;
  labelFontColor?: string;
  labelFontName?: string;

  // Extended properties
  id?: string;
  name?: string;
  additionalElements?: PPTElement[];
  accentImageShape?: 'default' | 'big' | 'mixed';
  card?: {
    enabled: boolean;
    borderRadius: number;
    borderWidth: number;
    fill: 'none' | 'full' | 'semi';
    shadow: PPTElementShadow;
    backgroundColor: string;
    textColor: string;
  };
}

export interface SlideTemplate {
  name: string;
  id: string;
  cover: string;
}